requirements.txt
google-cloud-texttospeech==2.14.1
tqdm==4.65.0
